+++ learn/src/collection/students/DateUtils.java
@@ -1,0 +1,13 @@
+package collection.students;
+
+import java.text.SimpleDateFormat;
+import java.util.Date;
+
+public class DateUtils {
+
+    private static final SimpleDateFormat formatter = new SimpleDateFormat("dd/MM/yyyy");
+
+    static String formatDate(Date date) {
+        return formatter.format(date);
+    }
+}
