+++ learn/src/core/oop/oop/principeoop/abstraction/1.md
