+++ learn/src/core/oop/oop/principeoop/incapsulation/1.md
@@ -1,0 +1,33 @@
+public final class core.oop.core.oop.immutable.ImmutableClassMy { 
+private final int myNumber; //инкапсулировали, скрыли и зафиналили переменную, ее не знают не могут изменить
+
+public core.oop.core.oop.immutable.ImmutableClassMy (int myNumber)
+{ 
+this.myNumber = myNumber; // конструктор, присваивающий аргумент к переменной класса 
+}
+
+public int getMyNumber() { 
+return myNumber; 
+} 
+}
+
+Пример кода с этим классом
+
+public class core.oop.core.oop.immutable.Main { public static void main(String[] args) { 
+
+ // Создаем объект core.oop.core.oop.immutable.ImmutableClassMy с значением myNumber = 42 
+ core.oop.core.oop.immutable.ImmutableClassMy immutableObj = new core.oop.core.oop.immutable.ImmutableClassMy(42);
+
+    // Получаем значение myNumber из объекта
+    int myNumber = immutableObj.getMyNumber();
+    System.out.println("myNumber = " + myNumber); //вывели 42
+
+    // Попытка изменить значение myNumber приведет к ошибке компиляции, так как переменная myNumber в классе core.oop.core.oop.immutable.ImmutableClassMy объявлена как final
+    // immutableObj.myNumber = 100; // Ошибка компиляции
+
+    // Создаем новый объект core.oop.core.oop.immutable.ImmutableClassMy с другим значением myNumber, будет успех
+    core.oop.core.oop.immutable.ImmutableClassMy anotherObj = new core.oop.core.oop.immutable.ImmutableClassMy(100);
+    int anotherNumber = anotherObj.getMyNumber();
+    System.out.println("anotherNumber = " + anotherNumber);
+}
+}
