+++ learn/src/core/oop/oop/principeoop/polymorphism/1.md
@@ -1,0 +1,38 @@
+Полиформизм - за одним интерфейсом могут скрываться разные реализации. Пример - человек может управлять машиной, значит может управлять и седаном, и хэтчбеком, и бмв, и мерседесом. Машины разные, реализация разная, но интерфейс один. Полиформизм позволяет единообразно обращаться к объектам различных классов (обычно имеющих одного родителя). Пример
+
+public class Animal {
+
+   public void voice() {
+
+       System.out.println("Голос!");
+   }
+}
+
+public class Dog extends Animal {
+
+
+   @Override
+   public void voice() {
+       System.out.println("Гав-гав!");
+   }
+}
+
+public class Cat extends Animal {
+
+   @Override
+   public void voice() {
+       System.out.println("Мяу!");
+   }
+}
+
+и 
+
+public class core.oop.core.oop.immutable.Main {
+
+   public static void main(String[] args) {
+
+       Animal dog = new Dog();
+       dog.voice();
+   }
+}
+Какой будет голос? Голос собаки)) тк мы переопределили метод аннотацией Override, а если нужна будет кошка, то создаем ссылку на объект кошки, new Cat. Родительский класс один, обращаемся одинаково, но поведение разное.
