+++ learn/src/core/trycatch/input/CountResultDto.java
@@ -1,0 +1,31 @@
+package core.trycatch.input;
+
+public class CountResultDto {
+    private int lineCount;
+    private int charCount;
+    private int wordCount;
+
+    public int getLineCount() {
+        return lineCount;
+    }
+
+    public void setLineCount(int lineCount) {
+        this.lineCount = lineCount;
+    }
+
+    public int getCharCount() {
+        return charCount;
+    }
+
+    public void setCharCount(int charCount) {
+        this.charCount = charCount;
+    }
+
+    public int getWordCount() {
+        return wordCount;
+    }
+
+    public void setWordCount(int wordCount) {
+        this.wordCount = wordCount;
+    }
+}
