--- learn/src/oop/incapsulation/1.md
@@ -1,33 +1,0 @@
-public final class immutable.ImmutableClassMy { 
-private final int myNumber; //инкапсулировали, скрыли и зафиналили переменную, ее не знают не могут изменить
-
-public immutable.ImmutableClassMy (int myNumber)
-{ 
-this.myNumber = myNumber; // конструктор, присваивающий аргумент к переменной класса 
-}
-
-public int getMyNumber() { 
-return myNumber; 
-} 
-}
-
-Пример кода с этим классом
-
-public class immutable.Main { public static void main(String[] args) { 
-
- // Создаем объект immutable.ImmutableClassMy с значением myNumber = 42 
- immutable.ImmutableClassMy immutableObj = new immutable.ImmutableClassMy(42);
-
-    // Получаем значение myNumber из объекта
-    int myNumber = immutableObj.getMyNumber();
-    System.out.println("myNumber = " + myNumber); //вывели 42
-
-    // Попытка изменить значение myNumber приведет к ошибке компиляции, так как переменная myNumber в классе immutable.ImmutableClassMy объявлена как final
-    // immutableObj.myNumber = 100; // Ошибка компиляции
-
-    // Создаем новый объект immutable.ImmutableClassMy с другим значением myNumber, будет успех
-    immutable.ImmutableClassMy anotherObj = new immutable.ImmutableClassMy(100);
-    int anotherNumber = anotherObj.getMyNumber();
-    System.out.println("anotherNumber = " + anotherNumber);
-}
-}
