--- learn/src/oop/inheritance/1.md
@@ -1,24 +1,0 @@
-Принцип наследования состоит в том, что новые классы можно создавать
-из уже существующих. При наследовании методы и поля существующеrо класса используются повторно (наследуются) вновь создаваемым классом, причем для адантации новоrо класса к новым условиям в неrо добавляют дополнительные поля и методы. Эrот прием иrрает в Java весьма важную роль. 
-
-Есть родительский класс, машины
-Там могут быть определенные переменные (пусть это будут общие атрибуты машины)
-И может быть дочерний класс, где могут быть свои атрибуты.
-
-Пример:
-public class Car{
-    int speed;
-    int weight;
-}
-
-public class oop.ElectricCar extends Car{
-    int electricEnginePower; //мощность электродвигателя
-}
-
-public class PetrolCar extends Car{
-    int petrolEnginePower; //мощность бензинового двигателя
-}
-
-Но есть проблема, суперкласс  может быть только один, согласно ООП. 
-Те нельзя сделать так - public class oop.ElectricCar extends Car, Veliche, будет ошибка компиляции.
-
